+++ solis-scraper/src/SolisScraper/Models/MqttConfiguration.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace SolisScraper.Models
+﻿namespace SolisScraper.Models
 {
     public class MqttConfiguration
     {
@@ -11,13 +9,7 @@ namespace SolisScraper.Models
         public string ClientId { get; set; } = "solis_scraper";
         public bool DebugLogging { get; set; }
         public string DiscoveryPrefix { get; set; } = "homeassistant";
-        public string NodeId { get; set; } = "solis";
-        public string UniqueIdPrefix { get; set; } = "solis_scraper";
-        public int FailureCap { get; set; } = 5;
-        
-        public TimeSpan IntervalError { get; set; } = TimeSpan.FromSeconds(30);
-        public TimeSpan IntervalZero { get; set; } = TimeSpan.FromSeconds(30);
-        public TimeSpan IntervalValue { get; set; } = TimeSpan.FromSeconds(15);
-        public TimeSpan IntervalDuplicateState { get; set; } = TimeSpan.FromMinutes(15);
+
+        public bool Dummy { get; set; }
     }
 }